requirements.txt
aiohttp>=3.10.10
python-dotenv>=1.0.1
yadisk>=3.1.0
httpx>=0.27.2
aiofiles>=24.1.0
psutil>=6.1.0